refactor: change tests packaging
